.github/actions/check-blocker-jira-issues/requirements.txt
atlassian-python-api==3.41.16
GitPython==3.1.44
requests==2.32.3
